feat: refactor add to list functionality with new components
