+++ presentation/src/main/java/com/madrid/presentation/screens/addtolist/CreateNewListItem.kt
@@ -1,0 +1,46 @@
+package com.madrid.presentation.screens.addtolist
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.unit.dp
+import com.madrid.designSystem.R
+import com.madrid.designSystem.component.MovioIcon
+import com.madrid.designSystem.component.MovioText
+import com.madrid.designSystem.theme.Theme
+
+@Composable
+fun CreateNewListItem(
+    onListCreated: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        modifier = modifier
+            .fillMaxWidth()
+            .clickable(
+                role = Role.Button,
+                onClickLabel = "Create a new list"
+            ) { onListCreated() }
+            .padding(horizontal = 16.dp, vertical = 32.dp),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        MovioText(
+            text = "Create a new list",
+            textStyle = Theme.textStyle.body.mediumMedium12,
+            color = Color.White,
+        )
+    }
+}
